Move custom service-specific functions to own package. This ensures that `make gen` succeeds.
