+++ aws/internal/keyvaluetags/custom/iam_tags.go
@@ -1,10 +1,11 @@
-package keyvaluetags
+package custom
 
 import (
 	"fmt"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/service/iam"
+	"github.com/terraform-providers/terraform-provider-aws/aws/internal/keyvaluetags"
 )
 
 // Custom IAM tag service update functions using the same format as generated code.
@@ -12,8 +13,8 @@ import (
 // IamRoleUpdateTags updates IAM role tags.
 // The identifier is the role name.
 func IamRoleUpdateTags(conn *iam.IAM, identifier string, oldTagsMap interface{}, newTagsMap interface{}) error {
-	oldTags := New(oldTagsMap)
-	newTags := New(newTagsMap)
+	oldTags := keyvaluetags.New(oldTagsMap)
+	newTags := keyvaluetags.New(newTagsMap)
 
 	if removedTags := oldTags.Removed(newTags); len(removedTags) > 0 {
 		input := &iam.UntagRoleInput{
@@ -47,8 +48,8 @@ func IamRoleUpdateTags(conn *iam.IAM, identifier string, oldTagsMap interface{},
 // IamUserUpdateTags updates IAM user tags.
 // The identifier is the user name.
 func IamUserUpdateTags(conn *iam.IAM, identifier string, oldTagsMap interface{}, newTagsMap interface{}) error {
-	oldTags := New(oldTagsMap)
-	newTags := New(newTagsMap)
+	oldTags := keyvaluetags.New(oldTagsMap)
+	newTags := keyvaluetags.New(newTagsMap)
 
 	if removedTags := oldTags.Removed(newTags); len(removedTags) > 0 {
 		input := &iam.UntagUserInput{
